requirements.txt
appnope==0.1.3
asgiref==3.5.2
asttokens==2.0.5
async-timeout==3.0.1
attrs==21.4.0
backcall==0.2.0
backports.entry-points-selectable==1.1.1
certifi==2022.6.15
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
decorator==5.1.1
distlib==0.3.4
Django==4.0.5
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.20.0
executing==0.8.3
factory-boy==3.2.1
Faker==13.14.0
filelock==3.7.1
flake8==4.0.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
ipython==8.4.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.6.1
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
pytest-django==4.5.2
pytest-flake8==1.1.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
tomli==2.0.1
tqdm==4.62.3
traitlets==5.3.0
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.10.0
virtualenv-clone==0.5.7
wcwidth==0.2.5
